Change dtype in OneHotEncoder

+++ encoder.py
@@ -68,7 +68,7 @@ class OneHotEncoder(skEncoder):
     Uses sklearn's OneHotEncoder and returns module sizes for use with fit method. For use with discrete features.
     """
     def __init__(self):
-        return super().__init__(sparse=False)
+        return super().__init__(sparse=False, dtype='int')
 
     def fit(self, X, y=None):
         super().fit(X)
